+++ Brokerages/Tradier/TradierBrokerageFactory.cs
@@ -19,6 +19,7 @@ using System.Globalization;
 using System.IO;
 using Newtonsoft.Json;
 using QuantConnect.Configuration;
+using QuantConnect.Data;
 using QuantConnect.Interfaces;
 using QuantConnect.Logging;
 using QuantConnect.Packets;
@@ -160,7 +161,11 @@ namespace QuantConnect.Brokerages.Tradier
             var issuedAt = Read<DateTime>(job.BrokerageData, "tradier-issued-at", errors);
             var lifeSpan = TimeSpan.FromSeconds(Read<double>(job.BrokerageData, "tradier-lifespan", errors));
 
-            var brokerage = new TradierBrokerage(algorithm.Transactions, algorithm.Portfolio, accountID);
+            var brokerage = new TradierBrokerage(
+                algorithm.Transactions, 
+                algorithm.Portfolio,
+                Composer.Instance.GetExportedValueByTypeName<IDataAggregator>(Config.Get("data-aggregator", "QuantConnect.Lean.Engine.DataFeeds.AggregationManager")),
+                accountID);
 
             // if we're running live locally we'll want to save any new tokens generated so that they can easily be retrieved
             if (Config.GetBool("tradier-save-tokens"))
